introduce new tests for void type

+++ src/compiler/normalizer/mod.rs
@@ -369,7 +369,7 @@ impl Normalizer {
                 self.type_resolver.hint_is(type_label, PrimitiveType::F64)?;
 
                 IRExpression::Constant {
-                    constant: IRConstant::Float(f64::from(x)),
+                    constant: IRConstant::Float(x),
                 }
             }
